backend/agent_core/requirements.txt
bedrock-agentcore
boto3
langgraph
langchain
langchain_aws
langgraph-checkpoint-aws
python-dotenv